+++ CHANGELOG.md
@@ -23,6 +23,13 @@
 - Enh #639: Simplify validation of JSON in `JsonHandler` using built-in PHP functions for PHP versions below 8.3
   (@arogachev)
 
+## 1.3.0 April 04, 2024
+
+- New #665: Add methods `addErrorWithFormatOnly()` and `addErrorWithoutPostProcessing()` to `Result` object (@vjik)
+- New #670, #677, #680: Add `Image` validation rule (@vjik, @arogachev)
+- New #678: Add `Date`, `DateTime` and `Time` validation rules (@vjik)
+- Enh #668: Clarify psalm types in `Result` (@vjik)
+
 ## 1.2.0 February 21, 2024
 
 - New #597, #608: Add debug collector for `yiisoft/yii-debug` (@xepozz, @vjik)